podcast-content-optimizer/backend/requirements.txt
Flask==3.0.0
Flask-CORS==5.0.0
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.7
google-generativeai==0.3.2
pydub==0.25.1
feedparser==6.0.11
tqdm==4.66.1
mutagen==1.47.0
Werkzeug==3.0.1
redis==5.0.8
